+++ R/zzz.R
@@ -1,0 +1,3 @@
+#' @importFrom sandwich vcovHC
+#' @importFrom pcse vcovPC
+NULL